requirements.txt
langchain==0.1.0
chromadb==0.4.22
google-generativeai==0.3.2
sentence-transformers==2.7.0
python-dotenv==1.0.0
requests==2.32.0
beautifulsoup4==4.12.0